Refactor filename fixing method in FilesystemAdapter
The commit includes the refactoring of the method used to fix filenames in the FilesystemAdapter class. This is mainly done to make the code more readable and efficient. Instead of using preg_match to find and modify the filename, pathinfo function is utilized to split the filename and extension before appending a timestamp.